requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.8
Django==5.1.1
django-filter==24.2
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
filelock==3.16.0
kombu==5.4.1
platformdirs==4.3.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.11.0
tzdata==2024.1
vine==5.1.0
virtualenv==20.26.4
wcwidth==0.2.13
